chore: Refactor all uses of config to use the config object from the config module
