+++ openadapt/models.py
@@ -9,7 +9,8 @@ from PIL import Image, ImageChops
 import numpy as np
 import sqlalchemy as sa
 
-from openadapt import config, window
+from openadapt import window
+from openadapt.config import config
 from openadapt.db import db
