+++ openadapt/record.py
@@ -33,7 +33,8 @@ with override_stdout_stderr():
 import mss.tools
 import psutil
 
-from openadapt import config, utils, video, window
+from openadapt import utils, video, window
+from openadapt.config import config
 from openadapt.db import crud
 from openadapt.extensions import synchronized_queue as sq
 from openadapt.models import ActionEvent
