+++ openadapt/scripts/scrub.py
@@ -34,7 +34,8 @@ with override_stdout_stderr():
 
 import numpy as np
 
-from openadapt import config, scrub, utils
+from openadapt import scrub, utils
+from openadapt.config import config
 
 
 def _make_frame(
