+++ tests/openadapt/privacy/test_providers.py
@@ -3,7 +3,7 @@
 import pytest
 import spacy
 
-from openadapt import config
+from openadapt.config import config
 
 if not spacy.util.is_package(config.SPACY_MODEL_NAME):  # pylint: disable=no-member
     pytestmark = pytest.mark.skip(reason="SpaCy model not installed!")
